Move locations of PPE pages
